pyproject.toml
[project]
name = "proyecto-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "psycopg>=3.3.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
]

[project.scripts]
dev = "fastapi dev ./main.py"

[tool.ruff]
line-length = 88
indent-width = 2
fix = true

[tool.ruff.format]
indent-style = "space"
